基本完成升级所需要的主要步骤

+++ trunk/core/lib/uFileUtil.pas
@@ -10,15 +10,54 @@ type
     class procedure DeleteDir(ADirName: string; ARecursive: Boolean = True); static;
     class function CreateFile(AFileName: string; AThenClose: Boolean = True): THandle; static;
     class function CreateDir(ADirName: string): Boolean; static;
+    class function CopyDir(ADirName: string; AToDir: string; nLx: Integer = 1): Boolean; static;
     class function GetRelativePath(ABasePath, AFile: string): string; static;
     class function GetAbsolutePathEx(ABasePath, ARelativePath: string): string; static;
   end;
 
 implementation
 
-uses Winapi.ShLwApi, Winapi.Windows, System.SysUtils, System.IOUtils;
+uses Winapi.ShLwApi, Winapi.Windows, System.SysUtils, System.IOUtils, Winapi.ShellAPI;
 
 { TFileUtil }
+
+class function TFileUtil.CopyDir(ADirName, AToDir: string; nLx:Integer): Boolean;
+Var
+  Opstruc: TshFileOpStruct;
+  frombuf, tobuf: Array [0 .. 128] of Char;
+begin
+  FillChar(frombuf, Sizeof(frombuf), 0);
+  FillChar(tobuf, Sizeof(tobuf), 0);
+  StrPcopy(frombuf, ADirName);
+  Case nLx of
+    1:
+      StrPcopy(tobuf, AToDir);
+  end;
+  With Opstruc Do
+  Begin
+    Wnd := 0;
+    Case nLx of
+      1:
+        wFunc := FO_COPY;
+      2:
+        wFunc := FO_DELETE;
+    Else
+      wFunc := FO_COPY;
+    end;
+    pFrom := @frombuf;
+    pTo := @tobuf;
+    fFlags := FOF_NOCONFIRMATION;
+    fAnyOperationsAborted := False;
+    hNameMappings := Nil;
+    lpszProgressTitle := Nil;
+  end;
+  try
+    ShFileOperation(Opstruc);
+    Result := True;
+  except
+    Result := False;
+  end;
+end;
 
 class function TFileUtil.CreateDir(ADirName: string): Boolean;
 begin
@@ -31,13 +70,15 @@ end;
 
 class procedure TFileUtil.DeleteDir(ADirName: string; ARecursive: Boolean = True);
 begin
-  TDirectory.Delete(ADirName, ARecursive);
+  if DirectoryExists(ADirName) then
+    TDirectory.Delete(ADirName, ARecursive);
 end;
 
 class function TFileUtil.CreateFile(AFileName: string; AThenClose: Boolean): THandle;
 begin
   Result := INVALID_HANDLE_VALUE;
-  if not TFileUtil.CreateDir(ExtractFileDir(AFileName)) then Exit;
+  if not TFileUtil.CreateDir(ExtractFileDir(AFileName)) then
+    Exit;
 
   Result := FileCreate(AFileName);
   if Result <> INVALID_HANDLE_VALUE then
@@ -47,30 +88,29 @@ begin
   end;
 end;
 
-
 class function TFileUtil.GetRelativePath(ABasePath, AFile: string): string;
   function GetAttr(IsDir: Boolean): DWORD;
   begin
     if IsDir then
-     Result := FILE_ATTRIBUTE_DIRECTORY
+      Result := FILE_ATTRIBUTE_DIRECTORY
     else
-     Result := FILE_ATTRIBUTE_NORMAL;
+      Result := FILE_ATTRIBUTE_NORMAL;
   end;
+
 var
-   p: array[0..MAX_PATH] of Char;
+  p: array [0 .. MAX_PATH] of Char;
 begin
-   PathRelativePathTo(p, PChar(ABasePath), GetAttr(False), PChar(AFile), GetAttr(True));
-   Result := StrPas(p);
+  PathRelativePathTo(p, PChar(ABasePath), GetAttr(False), PChar(AFile), GetAttr(True));
+  Result := StrPas(p);
 end;
 
-
-class function TFileUtil.GetAbsolutePathEx(ABasePath, ARelativePath:string):string;
+class function TFileUtil.GetAbsolutePathEx(ABasePath, ARelativePath: string): string;
 var
-   Dest:array [0..MAX_PATH] of char;
+  Dest: array [0 .. MAX_PATH] of Char;
 begin
-   FillChar(Dest, MAX_PATH+1,0);
-   PathCombine(Dest, PChar(ABasePath), PChar(ARelativePath));
-   Result := string(Dest);
+  FillChar(Dest, MAX_PATH + 1, 0);
+  PathCombine(Dest, PChar(ABasePath), PChar(ARelativePath));
+  Result := string(Dest);
 end;
 
 end.
